requirements.txt
torch
transformers
pillow
accelerate
bitsandbytes
streamlit
pandas